Move D2 UnicodeChar files
