+++ SlashGaming-Diablo-II-API/src/c/game_struct/c_d2_unicode_char.cc
@@ -38,7 +38,7 @@
 #include "../../../include/c/game_struct/d2_unicode_char.h"
 
 #include "../../../include/cxx/game_struct/d2_unicode_char.hpp"
-#include "../../cxx/game_struct/d2_unicode_char_impl.hpp"
+#include "../../cxx/game_struct/d2_unicode_char/d2_unicode_char_impl.hpp"
 
 struct D2_UnicodeChar* D2_UnicodeChar_CreateDefault(void) {
   return D2_UnicodeChar_CreateWithChar('\0');
